requirements.txt
google-generativeai>=0.4.0
pandas>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
PyPDF2>=3.0.0
opencv-python>=4.8.0
streamlit>=1.30.0
